requirements.txt
Django==5.0.6
djangorestframework==3.15
psycopg2-binary==2.9.9
django-filter==24.2
psycopg2>=2.9,<3
ipython==9.8.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.8.1
